refactor(core): move payload coercion out of lib
