+++ polylogue/archive/conversation/extraction.py
@@ -16,7 +16,7 @@ from polylogue.archive.semantic.facts import (
     build_conversation_semantic_facts,
 )
 from polylogue.archive.session.documents import WorkEventDocument
-from polylogue.lib.payload_coercion import (
+from polylogue.core.payload_coercion import (
     coerce_float,
     coerce_int,
     optional_date,
