+++ polylogue/archive/conversation/threads.py
@@ -10,7 +10,7 @@ from typing import TypeAlias, cast
 
 from polylogue.archive.session.documents import WorkThreadDocument, WorkThreadMemberEvidenceDocument
 from polylogue.archive.session.session_profile import SessionProfile
-from polylogue.lib.payload_coercion import (
+from polylogue.core.payload_coercion import (
     coerce_float,
     coerce_int,
     mapping_sequence,
