+++ polylogue/archive/session/models.py
@@ -14,7 +14,7 @@ from polylogue.archive.session.documents import (
     SessionPhaseDocument,
     SessionProfileDocument,
 )
-from polylogue.lib.payload_coercion import (
+from polylogue.core.payload_coercion import (
     coerce_float,
     coerce_int,
     int_pair,
